+++ index.js
@@ -15,14 +15,14 @@ export default class SlideDownPanel extends Component {
     super(props);
 
     const {
-      offsetTop, handlerHeight, containerMaximumHeight, containerBackgroundColor,
+      offsetTop, handlerHeight, initialHeight, containerMaximumHeight, containerBackgroundColor,
       containerOpacity, handlerDefaultView, handlerBackgroundColor, handlerOpacity
     } = props;
 
     this.state = {
       offsetTop: offsetTop != undefined ? offsetTop : DEFAULT_CONTAINER_HEIGHT,
       handlerHeight : handlerHeight != undefined ? handlerHeight: DEFAULT_CONTAINER_HEIGHT,
-      containerHeight : handlerHeight != undefined ? handlerHeight: DEFAULT_CONTAINER_HEIGHT,
+      containerHeight : handlerHeight != undefined ? initialHeight: handlerHeight,
       containerMinimumHeight : handlerHeight != undefined ? handlerHeight: DEFAULT_CONTAINER_HEIGHT,
       containerMaximumHeight : containerMaximumHeight != undefined ? containerMaximumHeight : height,
       containerBackgroundColor : containerBackgroundColor != undefined ? containerBackgroundColor : 'white',
@@ -32,6 +32,9 @@ export default class SlideDownPanel extends Component {
       handlerOpacity : handlerOpacity != undefined ? handlerOpacity : 1,
       isPanMoving: false
     };
+
+    this.hasLoaded = false;
+    this.previousContainerHeight = initialHeight;
   }
 
   componentWillMount() {
@@ -46,6 +49,7 @@ export default class SlideDownPanel extends Component {
   }
 
   componentDidMount() {
+    this.hasLoaded = true;
     // Make sure that handlerView is set
     if (this.state.handlerView == undefined) {
       throw "Set a handler view. Hint: It is a React Class."
@@ -53,43 +57,29 @@ export default class SlideDownPanel extends Component {
   }
 
   render() {
-    // Can't get to use below refactoring to work. Need to refactor!
-    // const styles = {
-    //   contatiner: {
-    //     position: 'absolute',
-    //     top: this.state.offsetTop,
-    //     opacity: this.state.containerOpacity,
-    //     paddingBottom: this.state.leastContainerHeight,
-    //     backgroundColor : this.state.containerBackgroundColor
-    //   },
-    //   handler: {
-    //     height : this.state.handlerHeight,
-    //     width : width,
-    //     justifyContent : 'center',
-    //     opacity : this.state.handlerOpacity,
-    //     backgroundColor : this.state.handlerBackgroundColor
-    //   }
-    // };
-
-
-    return (
-      this.state.isPanMoving ?
-      <View style={{
+    const styles = {
+      container: {
         position: 'absolute',
+        overflow: 'hidden',
         top: this.state.offsetTop,
         opacity: this.state.containerOpacity,
-        paddingBottom: this.state.leastContainerHeight,
         backgroundColor : this.state.containerBackgroundColor,
         height: this.state.containerHeight
-      }}>
+      },
+      handler: {
+        height : this.state.handlerHeight,
+        width : width,
+        justifyContent : 'center',
+        opacity : this.state.handlerOpacity,
+        backgroundColor : this.state.handlerBackgroundColor
+      }
+    };
+
+    return (
+      this.state.isPanMoving || !this.hasLoaded ?
+      <View style={styles.container}>
         {this.props.children}
-        <View style={{
-          height : this.state.handlerHeight,
-          width : width,
-          justifyContent : 'center',
-          opacity : this.state.handlerOpacity,
-          backgroundColor : this.state.handlerBackgroundColor
-        }} {...this.panResponder.panHandlers}>
+        <View style={styles.handler} {...this.panResponder.panHandlers}>
           {this.state.handlerView}
         </View>
       </View>
@@ -97,21 +87,9 @@ export default class SlideDownPanel extends Component {
       <Motion defaultStyle={{y: this.previousContainerHeight}} style={{y: spring(this.state.containerHeight, { stiffness: 200, damping: 30 })}}>
         {
           ({y}) => (
-            <View style={[{
-              position: 'absolute',
-              top: this.state.offsetTop,
-              opacity: this.state.containerOpacity,
-              paddingBottom: this.state.leastContainerHeight,
-              backgroundColor : this.state.containerBackgroundColor
-            }, { height: y}]}>
+            <View style={[styles.container, { height: y}]}>
               {this.props.children}
-              <View style={{
-                height : this.state.handlerHeight,
-                width : width,
-                justifyContent : 'center',
-                opacity : this.state.handlerOpacity,
-                backgroundColor : this.state.handlerBackgroundColor
-              }} {...this.panResponder.panHandlers}>
+              <View style={styles.handler} {...this.panResponder.panHandlers}>
                 {this.state.handlerView}
               </View>
             </View>