+++ CHANGELOG.md
@@ -9,3 +9,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.4] - 2016-04-01
 ### Changed
 - When moving slider, it does not animate. Only animates when touch ends. This improves responsiveness/ UX.
+
+## [1.0.5] - 2016-04-03
+### Added
+- Add new initialHeight props which can be used to set initial height.
+
+### Changed
+- Updated examples.
+- Clean up code.
+
+### Bug Fix
+- Fix IOS not hiding overflow properly.